feat/rn/reorganize routes to account for auth gated routes
