+++ apps/native/app/(app)/(home)/_layout.tsx
@@ -1,6 +1,6 @@
 import { Slot } from 'expo-router';
 
-import { HeaderWithBackNavigation } from '../../src/app-core/components/HeaderWithBackNavigation';
+import { HeaderWithBackNavigation } from '../../../src/app-core/components/headers/HeaderWithBackNavigation';
 
 // eslint-disable-next-line import/no-default-export
 export default function HomeLayout() {
