+++ apps/native/app/(app)/(home)/index.tsx
@@ -1,4 +1,4 @@
-import { HomeScreen } from '../../src/screens/HomeScreen';
+import { HomeScreen } from '../../../src/screens/HomeScreen';
 
 // eslint-disable-next-line import/no-default-export
 export default function Home() {
