+++ apps/native/app/(app)/examples/index.tsx
@@ -1,4 +1,4 @@
-import { ExamplesScreen } from '../../src/screens/ExamplesScreen';
+import { ExamplesScreen } from '../../../src/screens/ExamplesScreen';
 
 // eslint-disable-next-line import/no-default-export
 export default function Examples() {
