+++ apps/native/app/_layout.tsx
@@ -3,6 +3,7 @@ import { GluestackUIProvider } from '@gluestack-ui/themed';
 import { Slot } from 'expo-router';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
+import { AuthProvider } from '../src/authentication/state/AuthProvider';
 import { GraphQLProvider } from '../src/graphql/ApolloProvider';
 
 // eslint-disable-next-line import/no-default-export
@@ -11,7 +12,9 @@ export default function AppProvidersWrapper() {
     <GluestackUIProvider config={config}>
       <GraphQLProvider>
         <SafeAreaView>
-          <Slot />
+          <AuthProvider>
+            <Slot />
+          </AuthProvider>
         </SafeAreaView>
       </GraphQLProvider>
     </GluestackUIProvider>
